# apps/appointment-api-service/openapi.yaml
openapi: "3.1.0"
info:
  title: Keyloop — Multi-Tenant Vehicle Service Appointment API
  version: "1.0.0"
  description: |
    REST API for a multi-tenant vehicle service appointment system.
    All protected endpoints require a Bearer JWT in the Authorization header.
    All tenant-scoped endpoints require the `x-tenant-id` header matching the JWT `tenant_id` claim.
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    XTenantId:
      name: x-tenant-id
      in: header
      required: true
      description: UUID of the target tenant. Must match the `tenant_id` claim in the JWT (unless isSuperAdmin).
      schema:
        type: string
        format: uuid
  schemas:
...
    CustomerRequest:
      type: object
      required: [firstName, lastName, email]
      properties:
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
        phone:
          type: string
          maxLength: 30
    CustomerResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # ── Vehicle ───────────────────────────────────────────────────────────────
...
    ServiceTypeRequest:
      type: object
      required: [name, estimatedDurationMinutes]
      properties:
        name:
          type: string
          maxLength: 255
        estimatedDurationMinutes:
          type: integer
          minimum: 1
    ServiceTypeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        name:
          type: string
        estimatedDurationMinutes:
          type: integer
        createdAt:
          type: string
          format: date-time

    # ── Technician ────────────────────────────────────────────────────────────
    TechnicianRequest:
      type: object
      required: [firstName, lastName, email]
      properties:
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
    TechnicianResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # ── TechnicianSkill ───────────────────────────────────────────────────────
    TechnicianSkillRequest:
      type: object
      required: [technicianId, serviceTypeId]
      properties:
        technicianId:
          type: string
          format: uuid
        serviceTypeId:
          type: string
          format: uuid
    TechnicianSkillResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        technicianId:
          type: string
          format: uuid
        serviceTypeId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time

    # ── ServiceBay ────────────────────────────────────────────────────────────
    ServiceBayRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 100
    ServiceBayResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        name:
          type: string
        createdAt:
          type: string
          format: date-time

    # ── Appointment ───────────────────────────────────────────────────────────
...
paths:
...
  /customers:
    parameters:
      - $ref: '#/components/parameters/XTenantId'
    get:
      tags: [Customers]
      summary: List customers
      security:
        - BearerAuth: []
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/CustomerResponse' } }
                  total: { type: integer }
                  page: { type: integer }
                  pageSize: { type: integer }
        "400": { description: Missing x-tenant-id header }
        "403": { description: x-tenant-id does not match JWT }
    post:
      tags: [Customers]
      summary: Create a customer
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        "409": { description: Email already exists in this tenant }
  /customers/{id}:
    parameters:
      - $ref: '#/components/parameters/XTenantId'
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Customers]
      summary: Get a customer
      security:
        - BearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        "404": { description: Customer not found }
    put:
      tags: [Customers]
      summary: Update a customer
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
    delete:
      tags: [Customers]
      summary: Soft-delete a customer
      security:
        - BearerAuth: []
      responses:
        "204": { description: Customer deleted }
        "409": { description: Customer has active appointments }

  # ── Vehicles ────────────────────────────────────────────────────────────────
...
  /service-types:
    parameters:
      - $ref: '#/components/parameters/XTenantId'
    get:
      tags: [ServiceTypes]
      summary: List service types
      security:
        - BearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/ServiceTypeResponse' } }
                  total: { type: integer }
                  page: { type: integer }
                  pageSize: { type: integer }
    post:
      tags: [ServiceTypes]
      summary: Create a service type
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceTypeRequest'
      responses:
        "201": { content: { application/json: { schema: { $ref: '#/components/schemas/ServiceTypeResponse' } } } }
        "409": { description: Name already exists in this tenant }
  /service-types/{id}:
    parameters:
      - $ref: '#/components/parameters/XTenantId'
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [ServiceTypes]
      summary: Get a service type
      security:
        - BearerAuth: []
      responses:
        "200": { content: { application/json: { schema: { $ref: '#/components/schemas/ServiceTypeResponse' } } } }
    put:
      tags: [ServiceTypes]
      summary: Update a service type
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceTypeRequest'
      responses:
        "200": { content: { application/json: { schema: { $ref: '#/components/schemas/ServiceTypeResponse' } } } }
    delete:
      tags: [ServiceTypes]
      summary: Soft-delete a service type
      security:
        - BearerAuth: []
      responses:
        "204": { description: Deleted }
        "409": { description: Referenced by active appointment }

  # ── Technicians ──────────────────────────────────────────────────────────────
  /technicians:
    parameters:
      - $ref: '#/components/parameters/XTenantId'
    get:
      tags: [Technicians]
      summary: List technicians
      security:
        - BearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/TechnicianResponse' } }
                  total: { type: integer }
                  page: { type: integer }
                  pageSize: { type: integer }
    post:
      tags: [Technicians]
      summary: Create a technician
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TechnicianRequest'
      responses:
        "201": { content: { application/json: { schema: { $ref: '#/components/schemas/TechnicianResponse' } } } }
        "409": { description: Email already exists in this tenant }
  /technicians/{id}:
    parameters:
      - $ref: '#/components/parameters/XTenantId'
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Technicians]
      summary: Get a technician
      security:
        - BearerAuth: []
      responses:
        "200": { content: { application/json: { schema: { $ref: '#/components/schemas/TechnicianResponse' } } } }
    put:
      tags: [Technicians]
      summary: Update a technician
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TechnicianRequest'
      responses:
        "200": { content: { application/json: { schema: { $ref: '#/components/schemas/TechnicianResponse' } } } }
    delete:
      tags: [Technicians]
      summary: Soft-delete a technician
      security:
        - BearerAuth: []
      responses:
        "204": { description: Deleted }

  # ── TechnicianSkills ─────────────────────────────────────────────────────────
  /technicians/{id}/skills:
    parameters:
      - $ref: '#/components/parameters/XTenantId'
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [TechnicianSkills]
      summary: Add a skill to a technician
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TechnicianSkillRequest'
      responses:
        "201": { content: { application/json: { schema: { $ref: '#/components/schemas/TechnicianSkillResponse' } } } }
        "409": { description: Association already exists }
        "422": { description: Technician or ServiceType does not belong to this tenant }
  /technicians/{id}/skills/{serviceTypeId}:
    parameters:
      - $ref: '#/components/parameters/XTenantId'
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: serviceTypeId, in: path, required: true, schema: { type: string, format: uuid } }
    delete:
      tags: [TechnicianSkills]
      summary: Remove a skill from a technician
      security:
        - BearerAuth: []
      responses:
        "204": { description: Association removed }
        "404": { description: Association not found }

  # ── ServiceBays ──────────────────────────────────────────────────────────────
  /service-bays:
    parameters:
      - $ref: '#/components/parameters/XTenantId'
    get:
      tags: [ServiceBays]
      summary: List service bays
      security:
        - BearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/ServiceBayResponse' } }
                  total: { type: integer }
                  page: { type: integer }
                  pageSize: { type: integer }
    post:
      tags: [ServiceBays]
      summary: Create a service bay
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceBayRequest'
      responses:
        "201": { content: { application/json: { schema: { $ref: '#/components/schemas/ServiceBayResponse' } } } }
        "409": { description: Name already exists in this tenant }
  /service-bays/{id}:
    parameters:
      - $ref: '#/components/parameters/XTenantId'
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [ServiceBays]
      summary: Get a service bay
      security:
        - BearerAuth: []
      responses:
        "200": { content: { application/json: { schema: { $ref: '#/components/schemas/ServiceBayResponse' } } } }
    put:
      tags: [ServiceBays]
      summary: Update a service bay
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceBayRequest'
      responses:
        "200": { content: { application/json: { schema: { $ref: '#/components/schemas/ServiceBayResponse' } } } }
    delete:
      tags: [ServiceBays]
      summary: Soft-delete a service bay
      security:
        - BearerAuth: []
      responses:
        "204": { description: Deleted }

  # ── Appointments ─────────────────────────────────────────────────────────────
...